Get Brazilian municipalities codes

brazil_municipality_code() returns a vector with codes of the Brazilian Institute of Geography and Statistics (IBGE) for Brazilian municipalities.

Note: This function requires an active internet connection and the geobr package to be installed.

Usage

brazil_municipality_code(
  municipality,
  state = NULL,
  year = as.numeric(substr(Sys.Date(), 1, 4)),
  names = TRUE,
  ...
)

Arguments

municipality

(optional) A character vector with the name of the municipalities. If NULL the function returns all municipalities (default: NULL).

state

(optional) A character vector with the name of the states (default: NULL).

year

(optional) An integerish number indicating the year of the data regarding the municipalities (default: Sys.Date() |> substr(1, 4) |> as.numeric()).

names

(optional) A logical flag indicating whether to return the names of the municipalities as names of the vector (default: TRUE).

...

(optional) Additional arguments passed to brazil_municipality().

Value

An integer vector with the IBGE codes of Brazilian municipalities.

Details

Data from this function is based on data from the Brazilian Institute of Geography and Statistics (IBGE) via the geobr R package.

The geobr package is produced by Brazil's Institute for Applied Economic Research (IPEA) and access the Brazilian Institute of Geography and Statistics (IBGE) data. You can see a list of all geobr datasets by running geobr::list_geobr().

See also

Examples

library(curl)

# \dontrun{
  if (has_internet()) {
    brazil_municipality_code(municipality = "Belém")
  }
#> ! The closest map year to 2026 is 2024. Using year 2024 instead.
#> Belém-PA Belém-PB Belém-AL 
#>  1501402  2501906  2700805 
# }

# \dontrun{
  if (has_internet()) {
    brazil_municipality_code(municipality = "Belém", names = FALSE)
  }
#> ! The closest map year to 2026 is 2024. Using year 2024 instead.
#> [1] 1501402 2501906 2700805
# }

# \dontrun{
  if (has_internet()) {
    brazil_municipality_code(municipality = "Belém", state = "Pará")
  }
#> ! The closest map year to 2026 is 2024. Using year 2024 instead.
#> Belém-PA 
#>  1501402 
# }

# \dontrun{
  if (has_internet()) {
    brazil_municipality_code(c("Rio de Janeiro", "São Paulo"))
  }
#> ! The closest map year to 2026 is 2024. Using year 2024 instead.
#> Rio de Janeiro-RJ      São Paulo-SP 
#>           3304557           3550308 
# }