Get Brazilian municipalities latitude

brazil_municipality_latitude() returns the latitude of Brazilian municipalities.

Note: This function requires an internet connection to work and the geobr or geocodebr package to be installed, depending on the chosen method for retrieving coordinates.

Usage

brazil_municipality_latitude(
  municipality_code,
  year = as.numeric(substr(Sys.Date(), 1, 4)),
  coords_method = "geobr",
  names = TRUE,
  ...
)

Arguments

municipality_code

An integerish vector with the IBGE codes of Brazilian municipalities. Use brazil_municipality_code() to obtain codes from municipality names and states.

year

(optional) An integerish number indicating the year of the data regarding the municipalities (default: Sys.Date() |> substr(1, 4) |> as.numeric()).

coords_method

(optional) A character string indicating the method to retrieve the latitude and longitude coordinates of the municipalities (default: "geobr"). Options are:

"geobr": Uses read_municipal_seat() from the geobr package to retrieve the coordinates.

"geocodebr": Uses the geocode() from the geocodebr package to retrieve the coordinates.

names

(optional) A logical flag indicating whether to return the names of the municipalities as names of the vector (default: TRUE).

...

(optional) Additional arguments passed to brazil_municipality().

Value

A numeric vector with the latitude of the Brazilian municipalities.

Details

Data from this function is based on data from the Brazilian Institute of Geography and Statistics (IBGE) via the geobr R package.

The geobr package is produced by Brazil's Institute for Applied Economic Research (IPEA) and access the Brazilian Institute of Geography and Statistics (IBGE) data. You can see a list of all geobr datasets by running geobr::list_geobr().

See also

Examples

library(httr2)

# \dontrun{
  if (is_online()) {
    brazil_municipality_latitude(3550308)
  }
#> ! The closest map year to 2026 is 2024. Using year 2024 instead.
#> Using year/date 2024
#> Problem connecting to data server. Please try again in a few minutes.
#> Error in dplyr::mutate(dplyr::rename(dplyr::select(dplyr::as_tibble(read_municipality(year = closest_geobr_year(year,     type = "municipality", verbose = FALSE), showProgress = FALSE,     cache = !force)), name_muni, code_muni, code_state, abbrev_state),     municipality = name_muni, municipality_code = code_muni,     state_code = code_state, federal_unit = abbrev_state), municipality = to_title_case_pt(municipality,     articles = TRUE, conjunctions = FALSE, oblique_pronouns = FALSE,     prepositions = FALSE, custom_rules = c(`(.)\\bE( )\\b` = "\\1e\\2",         `(.)\\bÀ(s)?\\b` = "\\1à\\2", `(.)\\bD(((a|o)(s)?)|(e))\\b` = "\\1d\\2",         `(.)\\bE(m)\\b` = "\\1e\\2", `(.)\\bN((a|o)(s)?)\\b` = "\\1n\\2",         `(.)\\bD(el)\\b` = "\\1d\\2")), municipality_code = as.integer(municipality_code),     state = brazil_state(federal_unit), state_code = as.integer(state_code),     region = brazil_region(federal_unit), region_code = brazil_region_code(region)): ℹ In argument: `municipality = to_title_case_pt(...)`.
#> Caused by error in `to_title_case_pt()`:
#> ! Assertion on 'x' failed: Must be of type 'character', not 'NULL'.
# }

# \dontrun{
  if (is_online()) {
    brazil_municipality_latitude(c(3550308, 3500204))
  }
#> ! The closest map year to 2026 is 2024. Using year 2024 instead.
#> Using year/date 2024
#> Problem connecting to data server. Please try again in a few minutes.
#> Error in dplyr::mutate(dplyr::rename(dplyr::select(dplyr::as_tibble(read_municipality(year = closest_geobr_year(year,     type = "municipality", verbose = FALSE), showProgress = FALSE,     cache = !force)), name_muni, code_muni, code_state, abbrev_state),     municipality = name_muni, municipality_code = code_muni,     state_code = code_state, federal_unit = abbrev_state), municipality = to_title_case_pt(municipality,     articles = TRUE, conjunctions = FALSE, oblique_pronouns = FALSE,     prepositions = FALSE, custom_rules = c(`(.)\\bE( )\\b` = "\\1e\\2",         `(.)\\bÀ(s)?\\b` = "\\1à\\2", `(.)\\bD(((a|o)(s)?)|(e))\\b` = "\\1d\\2",         `(.)\\bE(m)\\b` = "\\1e\\2", `(.)\\bN((a|o)(s)?)\\b` = "\\1n\\2",         `(.)\\bD(el)\\b` = "\\1d\\2")), municipality_code = as.integer(municipality_code),     state = brazil_state(federal_unit), state_code = as.integer(state_code),     region = brazil_region(federal_unit), region_code = brazil_region_code(region)): ℹ In argument: `municipality = to_title_case_pt(...)`.
#> Caused by error in `to_title_case_pt()`:
#> ! Assertion on 'x' failed: Must be of type 'character', not 'NULL'.
# }

# \dontrun{
  if (is_online()) {
    brazil_municipality_latitude(c(3550308, 1000, 3500204))
  }
#> ! The closest map year to 2026 is 2024. Using year 2024 instead.
#> Using year/date 2024
#> Problem connecting to data server. Please try again in a few minutes.
#> Error in dplyr::mutate(dplyr::rename(dplyr::select(dplyr::as_tibble(read_municipality(year = closest_geobr_year(year,     type = "municipality", verbose = FALSE), showProgress = FALSE,     cache = !force)), name_muni, code_muni, code_state, abbrev_state),     municipality = name_muni, municipality_code = code_muni,     state_code = code_state, federal_unit = abbrev_state), municipality = to_title_case_pt(municipality,     articles = TRUE, conjunctions = FALSE, oblique_pronouns = FALSE,     prepositions = FALSE, custom_rules = c(`(.)\\bE( )\\b` = "\\1e\\2",         `(.)\\bÀ(s)?\\b` = "\\1à\\2", `(.)\\bD(((a|o)(s)?)|(e))\\b` = "\\1d\\2",         `(.)\\bE(m)\\b` = "\\1e\\2", `(.)\\bN((a|o)(s)?)\\b` = "\\1n\\2",         `(.)\\bD(el)\\b` = "\\1d\\2")), municipality_code = as.integer(municipality_code),     state = brazil_state(federal_unit), state_code = as.integer(state_code),     region = brazil_region(federal_unit), region_code = brazil_region_code(region)): ℹ In argument: `municipality = to_title_case_pt(...)`.
#> Caused by error in `to_title_case_pt()`:
#> ! Assertion on 'x' failed: Must be of type 'character', not 'NULL'.
# }

# \dontrun{
  if (is_online()) {
    brazil_municipality_latitude(c(3550308, NA, 3500204))
  }
#> ! The closest map year to 2026 is 2024. Using year 2024 instead.
#> Using year/date 2024
#> Problem connecting to data server. Please try again in a few minutes.
#> Error in dplyr::mutate(dplyr::rename(dplyr::select(dplyr::as_tibble(read_municipality(year = closest_geobr_year(year,     type = "municipality", verbose = FALSE), showProgress = FALSE,     cache = !force)), name_muni, code_muni, code_state, abbrev_state),     municipality = name_muni, municipality_code = code_muni,     state_code = code_state, federal_unit = abbrev_state), municipality = to_title_case_pt(municipality,     articles = TRUE, conjunctions = FALSE, oblique_pronouns = FALSE,     prepositions = FALSE, custom_rules = c(`(.)\\bE( )\\b` = "\\1e\\2",         `(.)\\bÀ(s)?\\b` = "\\1à\\2", `(.)\\bD(((a|o)(s)?)|(e))\\b` = "\\1d\\2",         `(.)\\bE(m)\\b` = "\\1e\\2", `(.)\\bN((a|o)(s)?)\\b` = "\\1n\\2",         `(.)\\bD(el)\\b` = "\\1d\\2")), municipality_code = as.integer(municipality_code),     state = brazil_state(federal_unit), state_code = as.integer(state_code),     region = brazil_region(federal_unit), region_code = brazil_region_code(region)): ℹ In argument: `municipality = to_title_case_pt(...)`.
#> Caused by error in `to_title_case_pt()`:
#> ! Assertion on 'x' failed: Must be of type 'character', not 'NULL'.
# }