Get Brazilian municipalities latitude

get_brazil_municipality_latitude() returns the latitude of Brazilian municipalities.

Usage

get_brazil_municipality_latitude(
  municipality_code,
  year = lubridate::year(Sys.Date()),
  coords_method = "geobr",
  names = TRUE,
  ...
)

Arguments

municipality_code

An integerish vector with the IBGE codes of Brazilian municipalities. Use get_brazil_municipality_code() to obtain codes from municipality names and states.

year

(optional) An integerish number indicating the year of the data regarding the municipalities (default: Sys.Date() |> lubridate::year()).

coords_method

(optional) A string indicating the method to retrieve the latitude and longitude coordinates of the municipalities. Options are:

"geobr": Uses read_municipal_seat() from the geobr package to retrieve the coordinates.

"geocodebr": Uses the geocode() from the geocodebr package to retrieve the coordinates. (default: "geobr").

names

(optional) A logical flag indicating whether to return the names of the municipalities as names of the vector (default: TRUE).

...

(optional) Additional arguments passed to get_brazil_municipality().

Value

A numeric vector with the latitude of the Brazilian municipalities.

Details

The data from this function is based on data from the Brazilian Institute of Geography and Statistics (IBGE) via the geobr R package.

The geobr package is produced by Brazil's Institute for Applied Economic Research (IPEA) and access the Brazilian Institute of Geography and Statistics (IBGE) data. You can see a list of all geobr datasets by running geobr::list_geobr().

See also

Examples

get_brazil_municipality_latitude(3550308)
#> São Paulo-SP 
#>    -23.56739 

get_brazil_municipality_latitude(c(3550308, 3500204))
#> São Paulo-SP    Adolfo-SP 
#>    -23.56739    -21.23273 

get_brazil_municipality_latitude(c(3550308, 1000, 3500204))
#> São Paulo-SP         <NA>    Adolfo-SP 
#>    -23.56739           NA    -21.23273 

get_brazil_municipality_latitude(c(3550308, NA, 3500204))
#> São Paulo-SP         <NA>    Adolfo-SP 
#>    -23.56739           NA    -21.23273