Get Brazilian state names by UTC

get_brazil_state_by_utc() returns a vector with the names of Brazilian states or abbreviations of Brazilian federal units by the UTC offset.

Usage

get_brazil_state_by_utc(utc = -3, type = "fu")

Arguments

utc

(Optional) An integerish number with the UTC offset. Available choices are -5, -4, -3, or -2 (Default: -3).

type

(Optional) A character string specifying the type of value to return. Available choices are "state" or "fu" (Default: "fu").

Value

A character vector with the names of Brazilian states or abbreviations of Brazilian federal units.

Details

The data from this function is based on the 2024b dataset (Released 2024-09-04) from the Internet Assigned Numbers Authority (IANA, 2024)

References

Internet Assigned Numbers Authority. (2024). Time zone database (No. 2024b) [Dataset]. https://www.iana.org/time-zones

See also

Examples

get_brazil_state_by_utc(-3, type = "fu")
#>  [1] "AL" "AP" "BA" "CE" "DF" "ES" "GO" "MA" "MG" "PA" "PB" "PR" "PE" "PI" "RJ"
#> [16] "RN" "RS" "SC" "SP" "SE" "TO"

get_brazil_state_by_utc(-3, type = "state")
#>  [1] "Alagoas"             "Amapá"               "Bahia"              
#>  [4] "Ceará"               "Distrito Federal"    "Espírito Santo"     
#>  [7] "Goiás"               "Maranhão"            "Minas Gerais"       
#> [10] "Pará"                "Paraíba"             "Paraná"             
#> [13] "Pernambuco"          "Piauí"               "Rio de Janeiro"     
#> [16] "Rio Grande do Norte" "Rio Grande do Sul"   "Santa Catarina"     
#> [19] "São Paulo"           "Sergipe"             "Tocantins"