Get paths to WorldClim data

get_wc_url() returns the URL(s) of a WorldClim data series.

get_wc_osf_id() returns the Open Science Framework (OSF) ID(s) for the LogoClim WorldClim data series backup.

Usage

get_wc_url(series, resolution = "all")

get_wc_osf_id(series, resolution = NULL)

Arguments

series

A character vector with the name of the WorldClim data series. The following options are available:

"hcd" or "historical-climate-data" or "historical climate data".

"hmwd" or "historical-monthly-weather-data" or "historical monthly weather data".

"fcd" or "future-climate-data" or "future climate data".

resolution

(optional) A character vector with the resolution of the WorldClim data series. The following options are available:

"all" (default) returns all available resolutions.

"10m" returns the 10m resolution.

"5m" returns the 5m resolution.

"2.5m" returns the 2.5m resolution.

"30s" returns the 30s resolution.

Value

A named character vector with the URL(s) or OSF ID(s) of the WorldClim data series.

See also

Examples

get_wc_url("hcd")
#>                       Historical climate data 
#> "https://worldclim.org/data/worldclim21.html" 

get_wc_osf_id("hcd")
#> Historical climate data 
#>                 "t2jfz" 

get_wc_url("historical climate data")
#>                       Historical climate data 
#> "https://worldclim.org/data/worldclim21.html" 

get_wc_url("hmwd")
#>              Historical monthly weather data 
#> "https://worldclim.org/data/monthlywth.html" 

get_wc_osf_id("hmwd")
#> Historical monthly weather data 
#>                         "rd3q5" 

get_wc_url("historical monthly weather data")
#>              Historical monthly weather data 
#> "https://worldclim.org/data/monthlywth.html" 

get_wc_url("fcd")
#>                              Future climate data (10m) 
#>  "https://worldclim.org/data/cmip6/cmip6_clim10m.html" 
#>                               Future climate data (5m) 
#>   "https://worldclim.org/data/cmip6/cmip6_clim5m.html" 
#>                             Future climate data (2.5m) 
#> "https://worldclim.org/data/cmip6/cmip6_clim2.5m.html" 
#>                              Future climate data (30s) 
#>  "https://worldclim.org/data/cmip6/cmip6_clim30s.html" 

get_wc_osf_id("fcd")
#>           Future climate data (10m)   Future climate data (5m | Part 1) 
#>                             "fz7gv"                             "fbgjh" 
#>   Future climate data (5m | Part 2) Future climate data (2.5m | Part 1) 
#>                             "76atf"                             "2fq8m" 
#> Future climate data (2.5m | Part 2) Future climate data (2.5m | Part 3) 
#>                             "8cqgp"                             "xgcve" 
#> Future climate data (2.5m | Part 4) Future climate data (2.5m | Part 5) 
#>                             "j4rvb"                             "q8azg" 
#> Future climate data (2.5m | Part 6)           Future climate data (30s) 
#>                             "g98z6"                                  NA 

get_wc_url("future climate data")
#>                              Future climate data (10m) 
#>  "https://worldclim.org/data/cmip6/cmip6_clim10m.html" 
#>                               Future climate data (5m) 
#>   "https://worldclim.org/data/cmip6/cmip6_clim5m.html" 
#>                             Future climate data (2.5m) 
#> "https://worldclim.org/data/cmip6/cmip6_clim2.5m.html" 
#>                              Future climate data (30s) 
#>  "https://worldclim.org/data/cmip6/cmip6_clim30s.html" 

get_wc_url("future climate data", "10m")
#>                             Future climate data (10m) 
#> "https://worldclim.org/data/cmip6/cmip6_clim10m.html" 

get_wc_url(c("hcd", "hmwd", "fcd"))
#>                                Historical climate data 
#>          "https://worldclim.org/data/worldclim21.html" 
#>                        Historical monthly weather data 
#>           "https://worldclim.org/data/monthlywth.html" 
#>                              Future climate data (10m) 
#>  "https://worldclim.org/data/cmip6/cmip6_clim10m.html" 
#>                               Future climate data (5m) 
#>   "https://worldclim.org/data/cmip6/cmip6_clim5m.html" 
#>                             Future climate data (2.5m) 
#> "https://worldclim.org/data/cmip6/cmip6_clim2.5m.html" 
#>                              Future climate data (30s) 
#>  "https://worldclim.org/data/cmip6/cmip6_clim30s.html" 

get_wc_osf_id(c("hcd", "hmwd", "fcd"))
#>             Historical climate data     Historical monthly weather data 
#>                             "t2jfz"                             "rd3q5" 
#>           Future climate data (10m)   Future climate data (5m | Part 1) 
#>                             "fz7gv"                             "fbgjh" 
#>   Future climate data (5m | Part 2) Future climate data (2.5m | Part 1) 
#>                             "76atf"                             "2fq8m" 
#> Future climate data (2.5m | Part 2) Future climate data (2.5m | Part 3) 
#>                             "8cqgp"                             "xgcve" 
#> Future climate data (2.5m | Part 4) Future climate data (2.5m | Part 5) 
#>                             "j4rvb"                             "q8azg" 
#> Future climate data (2.5m | Part 6)           Future climate data (30s) 
#>                             "g98z6"                                  NA 

get_wc_url(c("hcd", "hmwd", "fcd"), "5m")
#>                              Historical climate data 
#>        "https://worldclim.org/data/worldclim21.html" 
#>                      Historical monthly weather data 
#>         "https://worldclim.org/data/monthlywth.html" 
#>                             Future climate data (5m) 
#> "https://worldclim.org/data/cmip6/cmip6_clim5m.html" 

get_wc_osf_id(c("hcd", "hmwd", "fcd"), "5m")
#>           Historical climate data   Historical monthly weather data 
#>                           "t2jfz"                           "rd3q5" 
#> Future climate data (5m | Part 1) Future climate data (5m | Part 2) 
#>                           "fbgjh"                           "76atf"