Shift and rotate a raster or a vector

shift_and_rotate() shifts a raster or vector by a specified horizontal distance and rotates the data around the dateline.

This function is particularly useful for working with rasters and vectors that span the dateline (e.g., the Russian territory).

Usage

shift_and_rotate(x, dx = -45, precision = 5, overlap_tolerance = 0.1)

Arguments

x

A SpatRaster or SpatVector object to be shifted and rotated.

dx

A numeric value indicating the amount of the horizontal shift in degrees. Positive values shift to the right, negative values shift to the left (default: -45).

precision

(optional) A numeric value specifying the number of decimal digits to use when rounding longitude and latitude coordinates (default: 5).

overlap_tolerance

(optional) A numeric value specifying the tolerance for overlapping geometries when combining vectors. This value controls the allowable error when merging overlapping geometries (default: 0.1).

Value

A object of the same class as x that has been shifted and rotated by the specified amount in degrees.

See also

Examples