Test geobr package connection with its server

test_geobr_connection() tests if the geobr package can successfully connect to its server and retrieve data.

Note: This function requires an active internet connection and the geobr package to be installed.

Usage

test_geobr_connection()

Value

A logical flag indicating if the connection was successful.

See also

Examples

# \dontrun{
  library(geobr)
  library(httr2)

  if (is_online()) {
    test_geobr_connection()
  }
#> [1] TRUE
# }