Get paths to WorldClim data

worldclim_url() returns the URL(s) of a WorldClim data series.

Usage

worldclim_url(series, resolution = "all")

Arguments

series

A character vector with the name of the WorldClim data series (default: "hcd"). The following options are available:

'hcd' (Historical Climate Data).

'hmwd' (Historical Monthly Weather Data).

'fcd' (Future Climate Data).

resolution

(optional) A character vector with the resolution of the WorldClim data series (default: "all"). The following options are available:

"all" returns all available resolutions.

"10m" returns the 10m resolution.

"5m" returns the 5m resolution.

"2.5m" returns the 2.5m resolution.

"30s" returns the 30s resolution.

Value

A named character vector with the URL(s) of the WorldClim data series.

See also

Examples

worldclim_url("hcd")
#>                       Historical climate data 
#> "https://worldclim.org/data/worldclim21.html" 

worldclim_url("hmwd")
#>              Historical monthly weather data 
#> "https://worldclim.org/data/monthlywth.html" 

worldclim_url("fcd")
#>                              Future climate data (10m) 
#>  "https://worldclim.org/data/cmip6/cmip6_clim10m.html" 
#>                               Future climate data (5m) 
#>   "https://worldclim.org/data/cmip6/cmip6_clim5m.html" 
#>                             Future climate data (2.5m) 
#> "https://worldclim.org/data/cmip6/cmip6_clim2.5m.html" 
#>                              Future climate data (30s) 
#>  "https://worldclim.org/data/cmip6/cmip6_clim30s.html" 

worldclim_url("fcd", "10m")
#>                             Future climate data (10m) 
#> "https://worldclim.org/data/cmip6/cmip6_clim10m.html" 

worldclim_url(c("hcd", "hmwd", "fcd"))
#>                                Historical climate data 
#>          "https://worldclim.org/data/worldclim21.html" 
#>                        Historical monthly weather data 
#>           "https://worldclim.org/data/monthlywth.html" 
#>                              Future climate data (10m) 
#>  "https://worldclim.org/data/cmip6/cmip6_clim10m.html" 
#>                               Future climate data (5m) 
#>   "https://worldclim.org/data/cmip6/cmip6_clim5m.html" 
#>                             Future climate data (2.5m) 
#> "https://worldclim.org/data/cmip6/cmip6_clim2.5m.html" 
#>                              Future climate data (30s) 
#>  "https://worldclim.org/data/cmip6/cmip6_clim30s.html" 

worldclim_url(c("hcd", "hmwd", "fcd"), "5m")
#>                              Historical climate data 
#>        "https://worldclim.org/data/worldclim21.html" 
#>                      Historical monthly weather data 
#>         "https://worldclim.org/data/monthlywth.html" 
#>                             Future climate data (5m) 
#> "https://worldclim.org/data/cmip6/cmip6_clim5m.html"